requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
dash==2.9.1
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.2.3
fonttools==4.39.2
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-core==2.3.2
google-cloud-storage==2.7.0
google-crc32c==1.5.0
google-resumable-media==2.4.1
googleapis-common-protos==1.58.0
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
mysql-connector==2.2.9
mysql-connector-python==8.0.32
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
plotly==5.13.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
randomcolor==0.4.4.6
requests==2.28.2
rsa==4.9
six==1.16.0
tenacity==8.2.2
tomli==2.0.1
urllib3==1.26.15
Werkzeug==2.2.3
